+++ src/server/__tests__/parseSources.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 import {
   _getRegexps as getRegexps,
-  _parseFunctionCalls as parseFunctionCalls,
+  _parseWithRegexps as parseWithRegexps,
   _parseReactIntl as parseReactIntl,
 } from '../parseSources';
 
@@ -9,44 +9,77 @@ describe('Source parser', () => {
   describe('Function parser', () => {
     it('should work in a simple case', () => {
       const keys = {};
-      const regexpFunctionNames = getRegexps(['FOO']);
-      parseFunctionCalls(keys, 'FILENAME', `
+      const regexps = getRegexps(['FOO']);
+      parseWithRegexps(keys, 'FILENAME', `
         const a = 3;
         const str = FOO('Example');
-      `, regexpFunctionNames);
+      `, regexps);
       expect(keys).toMatchSnapshot();
     });
 
     it('should work with single and double quotes', () => {
       const keys = {};
-      const regexpFunctionNames = getRegexps(['FOO']);
-      parseFunctionCalls(keys, 'FILENAME', `
+      const regexps = getRegexps(['FOO']);
+      parseWithRegexps(keys, 'FILENAME', `
         const a = 3;
         const str = FOO('Example');
         const str2 = FOO("Example2");
-      `, regexpFunctionNames);
+      `, regexps);
       expect(keys).toMatchSnapshot();
     });
 
     it('should extract context if available', () => {
       const keys = {};
-      const regexpFunctionNames = getRegexps(['FOO']);
-      parseFunctionCalls(keys, 'FILENAME', `
+      const regexps = getRegexps(['FOO']);
+      parseWithRegexps(keys, 'FILENAME', `
         const a = 3;
         const str = FOO('context_Example');
-      `, regexpFunctionNames);
+      `, regexps);
       expect(keys).toMatchSnapshot();
     });
 
     it('should work with multiline strings', () => {
       const keys = {};
-      const regexpFunctionNames = getRegexps(['FOO']);
-      parseFunctionCalls(keys, 'FILENAME', `
+      const regexps = getRegexps(['FOO']);
+      parseWithRegexps(keys, 'FILENAME', `
         const a = 3;
         const str = FOO('context_This is a very
 long
 string');
-      `, regexpFunctionNames);
+      `, regexps);
+      expect(keys).toMatchSnapshot();
+    });
+
+    it('should work with specially-named functions', () => {
+      const keys = {};
+      const regexps = getRegexps(['$t']);
+      parseWithRegexps(keys, 'FILENAME', `
+        $t('hi') $t ('ho') $t( "there")
+      `, regexps);
+      expect(keys).toMatchSnapshot();
+    });
+  });
+
+  describe('Custom regexp parsing', () => {
+    it('should work with custom regexps', () => {
+      const keys = {};
+      const regexps = getRegexps([], ['{{\\s*(.*?)\\s*}}']);
+      parseWithRegexps(keys, 'FILENAME', `
+        whatever
+        {{ hey }}
+        whatever {{ho}} {{   let's go }} and {{ !! }}
+      `, regexps);
+      expect(keys).toMatchSnapshot();
+    });
+
+    it('should work with custom regexps and multiline strings', () => {
+      const keys = {};
+      const regexps = getRegexps([], ['===(.*?)===']);
+      parseWithRegexps(keys, 'FILENAME', `
+        ===A very
+long
+string===
+      `, regexps);
       expect(keys).toMatchSnapshot();
     });
   });