+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - If Mady finds the babel-core and babel-plugin-react-intl modules (included as `peerDependencies`), it will automatically enable the React Intl integration and extract messages from your `FormattedMessage`/`FormattedHTMLMessage` components and `defineMessages()` calls, as supported by the Babel plugin.
   - Add support for React Intl's `description` field, used to give more context to the translator than the message prefix (e.g. `someContext_A message`)
   - Translation compilation now generates `${lang}.reactIntl.json` files that you can use to bootstrap React Intl
+* [M] **Add custom regexp parsing**
 * **(Minor) breaking change**: Mady no longer looks for a `.madyrc` file at your project root. This file only contained two parameters (the rest of Mady's configuration being stored in the `<locales>/config.json` file): a port number (which for Mady's purposes is not so important) and the relative path to the locales folder. The relative path should now be passed in with the CLI `--dir` argument, instead of via `.madyrc`.
 
 ## 1.6.1 (August 25, 2016)